+++ lib/crypto-browser.js
