+++ lib/factory.js
@@ -2,7 +2,7 @@
  * Copyright (c) 2023 Digital Bazaar, Inc. All rights reserved.
  */
 import {ALGORITHM, ECDSA_CURVE, ECDSA_HASH} from './constants.js';
-import {webcrypto} from './ecdsa.js';
+import {webcrypto} from './crypto.js';
 
 // exposes sign method
 export function createSigner({id, secretKey}) {
