+++ lib/index.js
@@ -7,7 +7,7 @@ import {
   EXTRACTABLE,
   MULTIKEY_CONTEXT_V1_URL
 } from './constants.js';
-import {CryptoKey, webcrypto} from './ecdsa.js';
+import {CryptoKey, webcrypto} from './crypto.js';
 import {createSigner, createVerifier} from './factory.js';
 import {exportKeyPair, importKeyPair} from './serialize.js';
 import {toMultikey} from './translators.js';
