+++ lib/serialize.js
@@ -10,7 +10,7 @@ import {
   MULTIBASE_BASE58_HEADER,
   MULTIKEY_CONTEXT_V1_URL
 } from './constants.js';
-import {webcrypto} from './ecdsa.js';
+import {webcrypto} from './crypto.js';
 import {
   getNamedCurve,
   getSecretKeySize,
