+++ package.json
@@ -11,7 +11,7 @@
   "type": "module",
   "exports": "./lib/index.js",
   "browser": {
-    "./lib/ecdsa.js": "./lib/ecdsa-browser.js"
+    "./lib/crypto.js": "./lib/crypto-browser.js"
   },
   "files": [
     "lib/**/*.js"
